requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
pytz==2022.7.1
razorpay==1.3.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
